requirements.txt
# AI Factory Benchmarking Framework Dependencies

# Core dependencies
PyYAML>=6.0
requests>=2.31.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Database clients
psycopg2-binary>=2.9.0  # PostgreSQL

# Object storage
boto3>=1.28.0  # S3-compatible storage
botocore>=1.31.0

# Vector databases
chromadb>=0.4.0  # ChromaDB
faiss-cpu>=1.7.4  # Faiss (use faiss-gpu for GPU support)
pymilvus>=2.3.0  # Milvus
weaviate-client>=3.24.0  # Weaviate

# Monitoring and metrics
prometheus-client>=0.18.0

# CLI enhancements
rich>=13.0.0  # Rich terminal output
click>=8.1.0  # Command-line interface
tqdm>=4.66.0  # Progress bars

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Optional: Advanced analytics
# scikit-learn>=1.3.0

